# openapi.yml
openapi: 3.0.3
info:
  title: GymLog API
  description: API para la aplicación GymLog
  version: 1.0.0
...
paths:
...
  /api/v1/auth/social-login:
    post:
      tags:
        - Auth
      operationId: socialLogin
      summary: Login con proveedores sociales
      description: Valida un token social y devuelve un JWT de la aplicación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocialLoginDTO'
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDTO'
        '401':
          description: Token social inválido
        '500':
          description: Error del servidor
  /api/v1/auth/me:
    get:
      tags:
        - Auth
      operationId: getCurrentUser
      summary: Obtener usuario actual
      description: Obtiene la información del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Información del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '401':
          description: Token JWT inválido

  # Endpoints de Gestión de Tenants
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SocialLoginDTO:
      type: object
      required:
        - provider
        - token
      properties:
        provider:
          type: string
          enum: [google, apple, facebook]
        token:
          type: string
          description: Token de identidad del proveedor social
    AuthResponseDTO:
      type: object
      required:
        - jwt
        - user
      properties:
        jwt:
          type: string
          description: JWT de la aplicación GymLog
        user:
          $ref: '#/components/schemas/UserDTO'
    UserDTO:
      type: object
      required:
        - id
        - name
        - email
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        profileImageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...